extract pure cron logic into config/cron.ts
